GH-7414: Moved `fuzzy-search` to common.
 - Made `filter` sync.
 - Fixed flaw in the range validation. Args were not used.

Closes #7414.
